Add files via upload
pyBrick-DNA is a software written in Python that assembles components for the construction of genetic circuits. pyBrick-DNA is a user-friendly environment where scientists can select genetic sequences extracted from a large database, or input custom ones, to build genetic assemblies. Each component is modularly fused to generate a ready-to-use single DNA fragment. The outcome is a DNA sequence, along with a graphical representation. Moreover, the sequence and its labels can be exported as a GenBank file allowing its use with other synthetic biology tools and databases.